feat(preview): say what the file is beside the picture of it
The format and size were only shown for a file nothing could draw, which
missed the case that prompted them: a PDF on a machine without PDFium, or
a truncated scan, claims a preview, falls back to a type icon at paint
time, and reads as "no file matched". Every resolved file carries the line
now, in the Details frame and the viewer alike.

`describe` moves into `preview` so both callers share one implementation.
In the Details frame it sits top-left, over the picture: the other two
edges are spoken for by the action buttons and a recording's transport.
In the viewer it goes under the filename, which at full screen is the only
other thing saying which file this is.
